Pytest_PyATS/PyATS_DEMO1/requirements.txt
pyyaml
pytest
pyats[full]

